Remove direct dependency on `bevy_lint` from `bevy_cli` (#30)
This PR makes `bevy_lint` optional when installing `bevy_cli`. This
solves a number of issue:

- Nightly is technically no longer required for the main CLI, though I'm
going to address that in a follow-up.
- If you made a change to the `bevy_lint` crate, you used to have to run
`cargo build -p bevy_cli --bin bevy_lint_driver` for them to be applied.

In order to do this, I:

- Made `bevy lint` a facade command that passes all arguments to a
`bevy_lint` binary, if it can be found.
- Added the `bevy_lint` binary to the `bevy_lint` crate, and moved
`bevy_lint_driver` to `bevy_lint`.

Now, nothing lint-related will be installed if `bevy_lint` isn't except
for the facade command.