Add `missing_reflect` lint (attempt 2) (#139)
Closes #54. See #137 for past attempt.

This adds a lint that lets users require all components, resources, and
events derive `Reflect`. `Reflect` is often used by tools such as
`bevy-inspector-egui` to view the state of the world while the game is
running, but these tools are far less useful if they do not have access
to reflection data.

This lint is off by default, and intended to be used to require
`Reflect` for types within specific modules. For example:

```rust
#[deny(bevy::missing_reflect)]
mod components {
    use bevy::prelude::*;

    #[derive(Component)]
    struct MyComponent;
}
```

Developers are free to require it throughout the entire crate, however.

```rust
#![deny(bevy::missing_reflect)]
```

This lint gives lots of information when it is raised, such as where the
offending trait was implemented and how to fix it:


![image](https://github.com/user-attachments/assets/e0c33639-d81d-44e6-b7da-433b9f61a61d)

Thanks to `clippy_utils`'s
[`suggest_item_with_attr()`](https://doc.rust-lang.org/nightly/nightly-rustc/clippy_utils/sugg/trait.DiagExt.html#tymethod.suggest_item_with_attr),
the suggestion is machine applicable. This means that `cargo fix` can
automatically fix this lint, making migrating a code base super easy.