Rework binary detection (#169)
# Objective

Rework the detection of binaries again, aligning it closer to how Cargo
works and enabling several new features / fixing several bugs.

- Support multiple binaries in a single package
- Support running examples (closes #69)
- Fix detection of `target` folder location (closes #145)

With this PR, we can now just run `bevy run --example breakout web` in
the Bevy repo and it works without any additional configuration!

# Solution

The code previously made the wrong assumption that we want to run
packages with a binary target. Instead, we want to run the binary
targets itself. A package can even have multiple of them.
Additionally, examples need to be considered as binary targets. The
`--example` argument now needs to be factored into the detection
algorithm and the different path for examples must be considered in the
bundling step.

## Note for Reviewers

The crux of the changes is contained in `run/mod.rs` in the
`select_run_binary` function.
This implements the new algorithm to select the correct binary to run.

In `serve.rs`, the `index.html` file now needs to be generated
dynamically, in order to pick the path to the correct binary. Because
the `index.html` needs to be a _static_ string, it needs to be leaked to
create a static reference.