New default compilation profiles for web commands (#199)
# Objective

Closes #197.

The `bevy build web` and `bevy run web` commands currently use the same
default profiles as the `cargo` commands, so either `dev` or `release`
if the `--release` flag is provided.
However, the demands of web apps can be vastly different than native
apps when it comes to compilation.
For example, it's a lot more effective to optimize for low binary size
on the web, because it significantly reduces the loading time of the
app.

This causes two problems with the current implementation:
- The defaults are bad. If the user doesn't know about the differences,
they will probably have high loading times for their apps or potentially
longer compile times.
- Fixing the problem results in boilerplate. Once you set up custom
profiles for a better configuration, you will have to provide the
`--profile={profile_name}` argument *every* time you execute `bevy run
web`... which is both a hassle and easy to forget.

# Solution

The web commands now use the `web` or `web-release` profiles by default.
This means we can provide better defaults, while still allowing the user
to fully customize the configuration (by just defining these profiles
themselves in `Cargo.toml`).
It also means the user doesn't have to pass the `--profile` flag to the
web commands, in most cases.

Of course, by default there won't be any definitions for the `web` and
`web-release` profiles and we don't want to edit the user's `Cargo.toml`
directly. So if we configured `--profile=web` we would get an error.
We circumvent this problem with [`--config` flag
overrides](https://doc.rust-lang.org/cargo/reference/config.html#command-line-overrides),
which essentially allow you to add additional entries to the
`Cargo.toml`.
For example, `--config profile.web.inherits="dev"` allows us to add the
`web` profile. It can be further customized with default settings in a
similar way.

I just set up very basic default configurations for the new profiles,
optimizing them will be a task for a future PR.