Linter: Support Bevy 0.15 (#191)
This updates the linter to support Bevy 0.15. As #138 has not yet been
implemented, Bevy 0.14 support is being dropped.[^0]

Overall, it was surprisingly simple. I bumped the version of Bevy we use
in testing, then fixed any errors. The UI tests are so rigid that even
the slightest quiver could cause them to fail, which in this case is a
good thing!

The one thing of note is that I'm keeping support for
`World::many_entities()` and `many_entities_mut()` in
`bevy::panicking_world_methods` even though they are deprecated. Since
`panicking_methods` is likely a lint that would be denied in projects, I
don't want panicking methods to slip by just because someone added
`#[allow(deprecated)]` while migration their project. Once these methods
are completely removed from Bevy, we'll remove them from the linter.

[^0]: I hope this won't be permanent, though. If all goes to plan,
support for 0.14 will be re-added before the next release of
`bevy_lint`!