Add linter section to CLI `README.md` (#244)
Closes #239!

The section is pretty sparse, especially because I intentionally omitted
the installation instructions (they're too complicated + too likely to
go out-of-date, in my opinion).

As an extra drive-by fix, I switched the `cli` code block attributes to
`sh`, which fixes syntax highlighting for me in VSCode. I can revert
this if it's not desired!