Add E2E tests for `bevy build` (#222)
# Objective

Closes #221.

Improve our ability to catch errors in the Bevy CLI by using end-to-end
tests, which install the CLI and execute it on a test repository.
This should help us prevent errors like #218 in the future.

# Solution

Add a new test suite for `bevy build` in the integration `tests` folder.
It defines some utilities for the tests and then installs the CLI and
executes `bevy build` commands on a test repository.

The test repository is currently a simple workspace with a singular Bevy
application.
The Bevy app is just the minimal default setup.
In the future, we can expand this or add more packages to test more
behavior.

The tests assert that the command exits successfully and that the
expected binaries are created.
It tests dev and release profiles as well as the `web` equivalents.

The binaries are cleaned up before the tests to ensure that the checks
work as intended, but other build artifacts are retained across runs to
speed up the tests.
In the future, we might have to adjust our caching strategy for the
integration tests, as the test repo uses a different target directory.

To avoid the tests conflicting with each other, I use `serial_test` to
prevent them running in parallel.

We also now need to install the Bevy dependencies on Linux in CI, I
adopted the strategy used by the Bevy repo.

> [!NOTE]
>
> Keep in mind that running the tests will now overwrite your installed
Bevy CLI version.

# Future work

- Improve CI caching of the E2E tests
- Add tests for `bevy build web --bundle`
- Find a way to test `bevy run` and `bevy run web` (they don't
terminate, so are harder to test)
- Find a better place to store the utility functions, so we can reuse
them across tests

---------

Co-authored-by: DAA <42379074+DaAlbrecht@users.noreply.github.com>