Refresh web page when server restarts (#256)
# Objective

Closes #234.

If you terminate `cargo run` and then run it again, your old application
will have closed and a new one with the newest state opens.

With `bevy run web`, if you terminate the command, only the dev server
terminates, the application will keep running.
If you re-run the command, the tab in the browser will still contain
your old code.

Instead, we want the page to refresh to ensure that the page is
up-to-date.

# Solution

The solution here is adopted from (an old version of) Trunk's auto
reloading, see <https://github.com/trunk-rs/trunk/pull/231>.

The basic idea is that we inject a script into the `index.html` to open
a websocket to the dev server.
When the websocket connection closes (i.e., the server has been shut
down), we periodically try to reconnect again and trigger a reload when
that happens.

Actually, before trying to connect the websocket, we first send a `PING`
request to the same URL. This prevents us spamming the console with
errors due to the websocket connection failing, I couldn't find a good
way to suppress them.

Note that the autoreload script is currently just injected when not in
bundling mode and if the user didn't specify a custom `index.html`.
This is because we really only want to inject this in the dev version,
not in any deployed version.
I plan to lift this limitation in a future PR, but it will require a bit
of restructuring and I didn't want to blow this up more than necessary.

# Testing

With the Bevy CLI installed from this branch:

1. On any Bevy project, use `bevy run web --open`.
2. Wait for the build to complete and the app to load in the browser.
3. (If you want, perform small changes to the code).
4. Shut down the server by terminating the command.
5. Run `bevy run web`.
6. Once the build has completed, the browser tab with your app should
automatically refresh!

# Future Work

- Expand auto reload support to bundled mode and for custom
`index.html`.
- Having a websocket connection between the client and server means that
we are able to send messages from the server to the client (once I
figure out how exactly to do that). This opens up a vast array of
possibilities. With the `reload` message that's already supported with
the current client implementation, we could implement a `bevy watch`
(#245). We could also add cold and eventually hot reloading on asset
changes (#189). We could also display build errors directly in the
browser window, which I believe the current Trunk version is doing.