Make errors more informative when cargo metadata fails (#260)
When the a call to `cargo metadata` fails the only output the user sees
is "Error: EOF while parsing a value at line 1 column 0" which doesn't
explain anything. This change adds additional context.

I have made the choice to try to print stderr if it can be decoded as
utf-8 because that is where I believe all the useful stuff will always
be. If decoding fails I print the debug format for the Output struct
which will at least have some information.

fixes: #242