Create specification for lint module docs in contributing guide (#276)
This PR adds a new section to the linter contributing guide: "Lint
Module Docs". This page describes how lint documentation should be
formatted, and gives suggestions on how to do so.


[Rendered](https://github.com/TheBevyFlock/bevy_cli/blob/lint-module-docs/bevy_lint/docs/reference/lint-module-docs.md)

Furthermore, I went through all of our existing lint docs and updated
them to follow this specification. It should be a net-improvement! :)

---------

Co-authored-by: DAA <42379074+DaAlbrecht@users.noreply.github.com>