Automatically add a sensible title to web page (#273)
# Objective

Partially related to #123.
Removes one regression from
<https://github.com/TheBevyFlock/bevy_new_2d/pull/312>.

When you currently run an app with `bevy run web`, the browser tab will
just be called "Bey App".
This is not very descriptive or nice, so it would be better if we could
choose a better default value.

# Solution

- Adjust our bundling pipeline to allow pre-processing of the
`index.html` in all cases
- If not provided, add a `<title>` with the name of the binary
    - E.g. `bevy_new_2d` becomes "Bevy New 2d"

As a follow-up to this PR we can also more easily apply the
post-processing with the dev-only options for the dev server.

# Testing

1. Install the CLI from this branch.
2. Run any Bevy project with `bevy run web`. It should show the name of
the binary in the title. Try for example `bevy run --example=breakout
web` in the bevy repository.
3. If you provide a custom `web/index.html` file and set a `<title>`, it
should not be overwritten.