switch to `tracing` (#282)
# Objective

For features like https://github.com/TheBevyFlock/bevy_cli/issues/275 a
dedicated logging crate would be helpful to implement `-v`/`--verbose`.

# Solution

I opted for `tracing` instead of `log` even though it's a bit overkill
for us at the moment. `tracing` seems to be the most flexible and is
also mostly used in `bevy`. But I don't really have a strong opinion on
this.

# Testing

```sh
❯ bevy run web
2025-02-24T21:49:19.817751Z  INFO bevy_cli::build: Compiling to WebAssembly...
    Finished `web` profile [unoptimized + debuginfo] target(s) in 0.11s
2025-02-24T21:49:19.970038Z  INFO bevy_cli::build: Bundling JavaScript bindings...
2025-02-24T21:49:22.969352Z  INFO bevy_cli::web::bundle: No custom `web` folder found, using defaults.
2025-02-24T21:49:22.969406Z  INFO bevy_cli::run: Open your app at <http://localhost:4000>!
```
```sh
❯ export RUST_LOG="bevy_cli=error"
❯ bevy run web
    Finished `web` profile [unoptimized + debuginfo] target(s) in 0.15s
```
Closes #236

---------

Co-authored-by: BD103 <59022059+BD103@users.noreply.github.com>