Guard web functionality behind a feature flag (#297)
# Objective

Closes #198.

A lot of the CLI's complexity and dependencies currently come from the
web functionality.
If you don't want to run your apps in the browser, you shouldn't have to
carry that baggage!
This is especially relevant for CI, where this build time can really add
up over time.

# Solution

Add a new `web` feature flag (enabled by default), which guards all the
functionality and dependencies related to the web.

I took this opportunity to move more web-only functions to the `web`
module.

# Testing

To prevent regressions on the feature combinations, we now check them
with clippy.