Log invocations of external commands (#290)
# Objective

One goal of the Bevy CLI is to be fully optional.
That means that a user should be able to fully replicate the behavior of
the CLI by external commands, even though it might be a lot less
user-friendly.

# Solution

1. Add a new flag `--verbose`, `-v` that sets the `tracing` level to
`debug`.
2. Create a wrapper type for `Command` that ensures all commands are
handled in a similar way (logging the name, and arguments of the
command)

# Testing 

cargo test and see:
https://github.com/TheBevyFlock/bevy_cli/pull/290#issuecomment-2699068479

Closes #275