Add CLI config in Cargo.toml to set feature flags (#331)
# Objective

Closes #250.

Bevy apps often want to configure different features in dev vs. release
mode or when targeting native vs. web.
With native `cargo`, that's currently a pain to do.

With this PR, the user can configure `features` and `default_features`
in the CLI in general, but also specific for the different modes as
mentioned above.

# Solution

The `cargo metadata` output also includes the contents of the `metadata`
table of each package.
We retrieve the value for the package that we build for and then extract
the config from that metadata.

For example, a user can configure
`package.metadata.bevy_cli.web.release` to set up features when
compiling for the web in release mode.
The `bevy_cli.release` and `bevy_cli.web` configs are hereby merged into
the more specific configs to make configuration a bit less boilerplaty.

Note that the workspace config is currently _not_ considered, because
configuring features only makes sense for packages.

# Testing

Can be tested with `bevy_new_2d`, see
<https://github.com/TheBevyFlock/bevy_new_2d/pull/357>.

# Feedback wanted

- Is the config merge order base -> profile -> platform ->
platform-profile the one you'd expect?