Add lint: unconventional_naming (#345)
Adds a new lint called `unconventional_naming` that checks whether a
bevy trait is implemented for a structure that does not follow the
commonly used naming pattern.

So far, this only includes:
- `SystemSet`: should end in `Set`
- `Plugin`: should end in `Plugin`

Additionally, the Lint `plugin_not_ending_in_plugin` Lint got removed as
this is now handled in the `unconventional_naming` lint.

Closes #183

---------

Co-authored-by: BD103 <59022059+BD103@users.noreply.github.com>