Improve README, pass 1 (#405)
In preparation for a v0.1.0 release of the CLI, we need to improve the
documentation.
This is a first pass to make the README more useful.

1. **Improved installation instructions**: Now also include easy copy &
paste instructions on how to install the linter. I know this is annoying
to maintain, as we need to manually update the toolchain and version
number when they change. But I think it's well worth it to make the
setup as easily as possible for the user.
2. **Quick start guide**: Simple commands to copy & paste to get
introduced to all main CLI commands.
3. **Improved structure**: The sections are now ordered in the way you
would use the CLI: You create a new project, check its code quality
while developing and then run it in the browser.

---------

Co-authored-by: BD103 <59022059+BD103@users.noreply.github.com>