Write guide on releasing the CLI (#447)
This proposes a process for releasing the CLI, based it off of the
linter release guide. Some noticeable points:

1. Releasing the CLI requires Bevy engine maintainer approval.
- This smooths migration for when the CLI becomes an official Bevy tool,
since it will become required then.
2. The Github release description looks very similar [to the
linter's](https://github.com/TheBevyFlock/bevy_cli/releases/tag/lint-v0.3.0).
This is good for consistency, but may be confusing and hard to tell the
difference between the two. Any ideas on how to differentiate them
further?