fix: rebrand to "the prototype Bevy CLI" and "Bevy CLI (Alpha)"
Although it is simpler for users if the CLI rebranded to a different name, such as "The Bevy Flock CLI", we're on a deadline to get the CLI shipped before the game jam at the end of the month. Marking the CLI as an alpha release, even though the technology itself is relatively stable, is the compromise we've come to.