CLI v0.1.0-alpha.1 post-release (#455)
This is a follow-up to #454. Now that the CLI has been released, we can
follow the
[post-release](https://thebevyflock.github.io/bevy_cli/contribute/cli/release.html#post-release)
steps to lift the feature-freeze.

I chose to make the development version v0.1.0-dev instead of
v0.1.0-alpha.2-dev, since the prior is shorter. Our next release will be
v0.1.0-alpha.2, but I think v0.1.0-alpha.2-dev is just too long.