Fix `cargo-binstall` for CLI (#458)
#455 revealed a potential bug in `cargo-binstall` where the `{ version
}` package URL template did not respect the `--version` flag (see
https://github.com/cargo-bins/cargo-binstall/issues/2165).

This PR implements a temporary hack that hard-codes the latest version
into the package URL instead of using `{ version }`. This fixes
`cargo-binstall` until a better solution is found.

Additionally, as per @cart's
[request](https://discord.com/channels/691052431525675048/1278871953721262090/1375561563863257179),
this PR minimizes the `cargo-binstall` installation instructions so they
are no longer the default, recommended way of doing things. We now
suggest `cargo install` by default, and mention that precompiled
binaries are available.

---------

Co-authored-by: Ben Frankel <ben.frankel7@gmail.com>