update `duplicate_bevy_dependencies` ui tests to bevy 0.16 (#469)
I wanted to update the `duplicate_bevy_dependencies` lint test for a
while now. While updating, I realized that `0.16` is not supported since
that's not a SemVer, so I adjusted the `version` parsing a bit. It's
still not ideal, and I think in the end the only good solution is to
manually implement a resolver so we can directly compare the
VersionReqs.

---------

Co-authored-by: BD103 <59022059+BD103@users.noreply.github.com>