Added a 'host' argument for 'bevy run web' (#472)
Added a host argument to allow a different address to be specified to
host the web server.
This allows for setting the address to something like 0.0.0.0 to allow
connections from all hosts. Or specifying the device's address on a VPN.

Also changed how we were declaring the address and URL to be displayed
to the user to allow for support of IPv6 addresses.

---------

Co-authored-by: BD103 <59022059+BD103@users.noreply.github.com>