Make `--no-default-features` be a toggle instead of option (#473)
Fixes #471.

The changes the behavior of the `--no-default-features` flag. Before it
required a value to be passed:

```sh
# Disable default features
bevy build --no-default-features true
# Do not disable default features
bevy build --no-default-features false
```

Now the flag is a toggle:

```sh
# Disable default features
bevy build --no-default-features
# Do not disable default features
bevy build
```

This is a breaking change, as passing `true` or `false` is no longer
recognized.