Allow customization of wasm-opt args (#486)
# Objective

Closes #442.

For further optimization, the user might want to get full control over
the flags that are passed to `wasm-opt`.
At the same time, users also need an easy way to enable or disable
`wasm-opt`, e.g. disabling it even in release mode or enabling it in dev
mode.

# Solution

I changed the ArgBuilder to deal with OsString instead of String, since
that's what we need to pass to the Command anyway.
This simplified this change, sorry for the noise in the diff.

- The `--wasm-opt` CLI flag now also accepts multiple string values.
These are passed as args to `wasm-opt`. Alternatively, you can still use
`true` to enable the default flags or `false` to disable `wasm-opt`
entirely.
- The `wasm-opt` config option now also accepts an array of args:

  ```toml
  [package.metadata.bevy_cli.web.release]
  wasm-opt = ["-Oz", "--enable-bulk-memory"]
  ```

# Testing

In <https://github.com/TimJentzsch/bevy_complex_repo>, the `workspace`
folder now contains a `wasm_opt` package where the new options are used
in the `Cargo.toml`.