Fix docs redirects (#500)
When we migrated to `mdbook`, I added redirect pages to send people to
the new page. For example,
<https://thebevyflock.github.io/bevy_cli/bevy_lint/> should redirect to
<https://thebevyflock.github.io/bevy_cli/linter/index.html>.

Unfortunately, I made these redirects absolute paths, which don't work
because everything in the website is prefixed with `/bevy_cli`. This PR
switches the redirects to relative paths, so they work both with `mdbook
serve` and when published through Github Pages.

While they're ugly, in a month we should be good to completely remove
most of these redirects.