Restructure commands (#498)
This PR places all available commands into a separate `commands` module.
This should make it easier to navigate the repository and to find the
entry points for all commands.

It also opens the way to further improve our documentation on docs.rs.

Note that this is a **breaking change** for users that depend on the
library of the CLI, as it affects some import paths.
I added a small section to the migration guide for this.

---------

Co-authored-by: BD103 <59022059+BD103@users.noreply.github.com>