Make linter action more ergonomic (#501)
# Context

The [Github Action we provide for the
linter](https://thebevyflock.github.io/bevy_cli/linter/github-actions.html)
isn't great. It works well for installing from a release or from `main`,
but breaks down when installing a specific commit. This is because we
currently hardcode values in `action.yml`:


https://github.com/TheBevyFlock/bevy_cli/blob/ff8d39aba29f6f04aecd7f8ec429d173649646ee/bevy_lint/action.yml#L18-L19


https://github.com/TheBevyFlock/bevy_cli/blob/ff8d39aba29f6f04aecd7f8ec429d173649646ee/bevy_lint/action.yml#L25-L31

This is an extra maintenance burden that we shouldn't have to do! If the
user wants to install a specific commit of the linter, it should work as
expected instead of installing the `main` branch:

```yml
# Should install this specific commit of the linter, but installs `main` instead.
- name: Install `bevy_lint`
  uses: TheBevyFlock/bevy_cli/bevy_lint@ff8d39a
```

# Solution

Thankfully, the fix is easier than expected. `action.yml` is a composite
action, meaning it's a single YAML file that slings around other actions
and shell scripts to do things. The great thing about this is that, when
an action runner downloads a composite action, it downloads **the entire
repository**. That means `rust-toolchain.toml` and the correct source
code for `bevy_lint` already exist, we just have to install them!

This PR does just that. It uses the venerable
[Taplo](https://taplo.tamasfe.dev/) to read `rust-toolchain.toml` and
then runs `cargo install --path ...` to install the linter. The action
especially uses the `${{ github.action_path }}` variable to locate the
`bevy_lint` folder.

# Testing


[`linter-action.yml`](https://github.com/TheBevyFlock/bevy_cli/blob/ff8d39aba29f6f04aecd7f8ec429d173649646ee/.github/workflows/linter-action.yml)
is doing a lot of the heavy-lifting here by testing the linter action in
CI. I'll also spin up a separate repository to prove it works for others
as well!