Use Regex instead of `taplo` to extract channel in linter action (#526)
Fixes #510.

This PR modifies `bevy_lint`'s action to use Regex to capture the
toolchain channel in `rust-toolchain.toml`. Specifically, it uses this
pattern:

```regex
^channel = "(.+)"$
```

I recommend [looking at this in
Regex101](https://regex101.com/r/zIQPCg/1) to understand what its doing.
In short, it looks for the line `channel = "..."` and captures
everything between the double quotes.

Additionally, I switched the toolchain components to be completely
hard-coded within the action. As these components are unlikely to ever
change, I figured it was far simpler than trying to use _more_ Regex to
extract them from `rust-toolchain.toml`. This can be changed in the
future!

---------

Co-authored-by: DAA <42379074+DaAlbrecht@users.noreply.github.com>