Rustup to `nightly-2025-08-07` (#539)
[Rust 1.89.0 was released
yesterday](https://blog.rust-lang.org/2025/08/07/Rust-1.89.0/), meaning
we can once again [update to a new nightly Rust
toolchain](https://thebevyflock.github.io/bevy_cli/contribute/linter/how-to/upgrade-rust-toolchain.html).

# Notable Changes

It's been a relatively mild release, at least in terms of the APIs we
depend on. The only thing visible is a adjustment in the alignment of
line numbers in diagnostics, which was caught by our UI tests.

# Drive-By Changes

I also fixed our docs, as before it had us updating the toolchain
_before_ we released a new version of `bevy_lint`. This caused
installation to fail, since the wrong toolchain would be used.

We now only change the toolchain upon releasing a new version of the
linter. (Except for the compatibility table, which we update now.)