Allow web headers to be defined in config (#544)
# Objective

Closes #538.

You currently can only specify additional web server headers via the
`--header` arg, but not persistently in the config.
This would be nice to simplify the command definition and make it easier
to share with others.

# Solution

You can now specify it in the config:

```toml
[package.metadata.bevy_cli]
headers = ["Cross-Origin-Opener-Policy:same-origin"]
```

Currently only in an array, in the future we might add map support as
well.

The headers from the config are applied _in addition to_ the headers
defined in the CLI.