fix: use owned strings and paths for artifact messages
JSON uses backslashes `\` to escape strings. In order for `serde_json` to convert the JSON `"C:\\path\\to\\file.exe"` to `"C:\path\to\file.exe"`, it must created a new owned `String`. Using zero-copy `&'a str` is not possible in this scenario, and causes deserialization to fail. This issue really only showed itself on Windows, because it uses backslashes as a path separator, but it could be a problem on other platforms.